backend/requirements.txt
Flask==3.0.3
pymongo==4.6.3
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
flask-cors==4.0.0
python-dotenv==1.0.1
